provotum/backend#5: Implement RESTful interface according to webservice specification in provotum/specification#2
