+++ src/main/java/org/provotum/backend/communication/socket/message/state/OpenVoteEventResponse.java
@@ -1,20 +1,20 @@
-package org.provotum.backend.communication.socket.message.event;
+package org.provotum.backend.communication.socket.message.state;
 
 import org.provotum.backend.communication.message.base.AResponse;
 import org.provotum.backend.communication.message.base.ResponseType;
 import org.provotum.backend.communication.message.base.Status;
 
-public class CloseVoteEventResponse extends AResponse {
+public class OpenVoteEventResponse extends AResponse {
 
     private String transaction;
-    private final ResponseType responseType = ResponseType.CLOSE_VOTE;
+    private final ResponseType responseType = ResponseType.OPEN_VOTE;
 
-    public CloseVoteEventResponse(String id, Status status, String message, String transaction) {
+    public OpenVoteEventResponse(String id, Status status, String message, String transaction) {
         super(id, status, message);
         this.transaction = transaction;
     }
 
-    public CloseVoteEventResponse(Status status, String message, String transaction) {
+    public OpenVoteEventResponse(Status status, String message, String transaction) {
         super(status, message);
         this.transaction = transaction;
     }
